Revert search improvements to simpler initial version
Because, while the improvement is supposed to allow for submitting a
search without a page reload, for some reason that is not working and
I lost the possibility of submitting a seach from a remote form.

Most probably both these problems can easily be solved with a little
tweaking of the js and html files. But that's currently not a priority.